EndreDokument.dokId fra Int til Long

+++ src/main/kotlin/no/nav/bidrag/dokument/dto/Journalpost.kt
@@ -83,7 +83,7 @@ data class RegistrereJournalpostCommand(
 @Schema(description = "Metadata for endring av et dokument")
 data class EndreDokument(
     @Schema(description = "Brevkoden til dokumentet") var brevkode: String? = null,
-    @Schema(description = "Identifikator av dokument informasjon") var dokId: Int = -1,
+    @Schema(description = "Identifikator av dokument informasjon") var dokId: Long = -1,
     @Schema(description = "Tittel på dokumentet") var tittel: String? = null
 )
